+++ src/main/kotlin/org/bytecamp/program_repair/astor_plugin/actions/AstorExecuteAstorAction.kt
@@ -5,12 +5,13 @@ import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.components.service
 import com.intellij.openapi.ui.Messages
 import org.bytecamp.program_repair.astor_plugin.services.AstorProjectService
+import kotlin.concurrent.thread
 
 class AstorExecuteAstorAction: AnAction() {
     override fun actionPerformed(e: AnActionEvent) {
         val service = e.project?.service<AstorProjectService>()!!
-        Messages.showMessageDialog(e.project, "Executing astor with arguments \n" + service.getConfig().toArgs().joinToString("\n"), "Astor", null)
-        val result = service.execute()
-        Messages.showMessageDialog(e.project, "Executed astor: $result", "Astor", null)
+        thread {
+            service.execute()
+        }
     }
 }